requirements.txt
fastapi
uvicorn
requests
pytest
psycopg2-binary
redis
httpx
